requirements.txt
appdirs==1.4.3
APScheduler==3.3.1
beautifulsoup4==4.5.3
click==6.7
cssselect==1.0.1
feedfinder2==0.0.4
feedparser==5.2.1
Flask==0.12
Flask-PyMongo==0.4.1
idna==2.5
itsdangerous==0.24
jieba3k==0.35.1
Jinja2==2.9.5
lxml==3.7.3
MarkupSafe==1.0
newspaper3k==0.1.9
nltk==3.2.2
olefile==0.44
packaging==16.8
Pillow==4.0.0
pymongo==3.4.0
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2016.10
PyYAML==3.12
requests==2.13.0
requests-file==1.4.1
six==1.10.0
tldextract==2.0.2
tzlocal==1.3
Werkzeug==0.12.1
